# docs/technical/03_api_specification.generated.yaml
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    OrphanCreate:
      properties:
        date_of_birth:
          format: date
          title: Date Of Birth
          type: string
        family_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Family Id
        family_name:
          maxLength: 100
          minLength: 1
          title: Family Name
          type: string
        father_death_date:
          anyOf:
          - format: date
            type: string
          - type: 'null'
          title: Father Death Date
        father_name:
          anyOf:
          - maxLength: 255
            type: string
          - type: 'null'
          title: Father Name
        first_name:
          maxLength: 100
          minLength: 1
          title: First Name
          type: string
        full_name_en:
          anyOf:
          - maxLength: 255
            type: string
          - type: 'null'
          title: Full Name En
        gender:
          enum:
          - M
          - F
          title: Gender
          type: string
        middle_name:
          anyOf:
          - maxLength: 100
            type: string
          - type: 'null'
          title: Middle Name
        nationality:
          anyOf:
          - maxLength: 2
            minLength: 2
            type: string
          - type: 'null'
          title: Nationality
        partner_organization_id:
          format: uuid
          title: Partner Organization Id
          type: string
      required:
      - first_name
      - family_name
      - date_of_birth
      - gender
      - partner_organization_id
      title: OrphanCreate
      type: object
...
    OrphanRead:
      properties:
        case_status:
          enum:
          - pending_review
          - approved
          - rejected
          - available
          - reserved
          - sponsored
          - graduated
          - deceased
          - archived
          title: Case Status
          type: string
        code:
          title: Code
          type: string
        created_at:
          format: date-time
          title: Created At
          type: string
        current_balance:
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Current Balance
          type: string
        date_of_birth:
          format: date
          title: Date Of Birth
          type: string
        family_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Family Id
        family_name:
          maxLength: 100
          minLength: 1
          title: Family Name
          type: string
        father_death_date:
          anyOf:
          - format: date
            type: string
          - type: 'null'
          title: Father Death Date
        father_name:
          anyOf:
          - maxLength: 255
            type: string
          - type: 'null'
          title: Father Name
        first_name:
          maxLength: 100
          minLength: 1
          title: First Name
          type: string
        full_name_en:
          anyOf:
          - maxLength: 255
            type: string
          - type: 'null'
          title: Full Name En
        gender:
          enum:
          - M
          - F
          title: Gender
          type: string
        id:
          format: uuid
          title: Id
          type: string
        is_sponsored:
          title: Is Sponsored
          type: boolean
        middle_name:
          anyOf:
          - maxLength: 100
            type: string
          - type: 'null'
          title: Middle Name
        nationality:
          anyOf:
          - maxLength: 2
            minLength: 2
            type: string
          - type: 'null'
          title: Nationality
        organization_id:
          format: uuid
          title: Organization Id
          type: string
        partner_organization_id:
          format: uuid
          title: Partner Organization Id
          type: string
        updated_at:
          format: date-time
          title: Updated At
          type: string
      required:
      - first_name
      - family_name
      - date_of_birth
      - gender
      - id
      - code
      - organization_id
      - partner_organization_id
      - family_id
      - case_status
      - is_sponsored
      - current_balance
      - created_at
      - updated_at
      title: OrphanRead
      type: object
    OrphanUpdate:
      description: 'Partial update. Every field optional — only the supplied keys are

        written. Status changes go through dedicated endpoints (workflow).'
      properties:
        date_of_birth:
          anyOf:
          - format: date
            type: string
          - type: 'null'
          title: Date Of Birth
        family_name:
          anyOf:
          - maxLength: 100
            minLength: 1
            type: string
          - type: 'null'
          title: Family Name
        father_death_date:
          anyOf:
          - format: date
            type: string
          - type: 'null'
          title: Father Death Date
        father_name:
          anyOf:
          - maxLength: 255
            type: string
          - type: 'null'
          title: Father Name
        first_name:
          anyOf:
          - maxLength: 100
            minLength: 1
            type: string
          - type: 'null'
          title: First Name
        full_name_en:
          anyOf:
          - maxLength: 255
            type: string
          - type: 'null'
          title: Full Name En
        gender:
          anyOf:
          - enum:
            - M
            - F
            type: string
          - type: 'null'
          title: Gender
        middle_name:
          anyOf:
          - maxLength: 100
            type: string
          - type: 'null'
          title: Middle Name
        nationality:
          anyOf:
          - maxLength: 2
            minLength: 2
            type: string
          - type: 'null'
          title: Nationality
      title: OrphanUpdate
      type: object
...
    Page_OrphanRead_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/OrphanRead'
          title: Items
          type: array
        limit:
          title: Limit
          type: integer
        offset:
          title: Offset
          type: integer
        total:
          title: Total
          type: integer
      required:
      - items
      - total
      - limit
      - offset
      title: Page[OrphanRead]
      type: object
...
    PublicOrphanCard:
      description: 'Curated public projection of an orphan. Every field here was

        chosen because it is safe to show an anonymous visitor.'
      properties:
        age_years:
          title: Age Years
          type: integer
        case_status:
          title: Case Status
          type: string
        code:
          title: Code
          type: string
        country:
          anyOf:
          - type: string
          - type: 'null'
          title: Country
        first_name:
          title: First Name
          type: string
        gender:
          enum:
          - M
          - F
          title: Gender
          type: string
        partner_organization_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Partner Organization Name
      required:
      - code
      - first_name
      - age_years
      - gender
      - country
      - case_status
      - partner_organization_name
      title: PublicOrphanCard
      type: object
    PublicOrphanDetail:
      description: Detail page — same fields as the card plus a short description.
      properties:
        age_years:
          title: Age Years
          type: integer
        case_status:
          title: Case Status
          type: string
        code:
          title: Code
          type: string
        country:
          anyOf:
          - type: string
          - type: 'null'
          title: Country
        first_name:
          title: First Name
          type: string
        gender:
          enum:
          - M
          - F
          title: Gender
          type: string
        partner_organization_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Partner Organization Name
        short_description:
          anyOf:
          - type: string
          - type: 'null'
          title: Short Description
      required:
      - code
      - first_name
      - age_years
      - gender
      - country
      - case_status
      - partner_organization_name
      title: PublicOrphanDetail
      type: object
    PublicPage:
      properties:
        items:
          items:
            $ref: '#/components/schemas/PublicOrphanCard'
          title: Items
          type: array
        limit:
          title: Limit
          type: integer
        offset:
          title: Offset
          type: integer
        total:
          title: Total
          type: integer
      required:
      - items
      - total
      - limit
      - offset
      title: PublicPage
      type: object
...
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
  securitySchemes:
    OAuth2PasswordBearer:
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login
      type: oauth2
info:
  title: Rufaqaa API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/v1/orphans:
    get:
      description: 'List orphans, optionally filtered by case_status and a search term.


        `q` does a Postgres full-text search against the trigger-maintained

        `search_vector` (covers Arabic + English name fields and the code) and

        also matches the code prefix directly so partial codes like "ORF-AB"

        still work.'
      operationId: list_orphans_api_v1_orphans_get
      parameters:
      - in: query
        name: limit
        required: false
        schema:
          default: 20
          maximum: 100
          minimum: 1
          title: Limit
          type: integer
      - in: query
        name: offset
        required: false
        schema:
          default: 0
          minimum: 0
          title: Offset
          type: integer
      - in: query
        name: case_status
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Case Status
      - in: query
        name: q
        required: false
        schema:
          anyOf:
          - maxLength: 100
            minLength: 1
            type: string
          - type: 'null'
          title: Q
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_OrphanRead_'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: List Orphans
      tags:
      - orphans
    post:
      operationId: create_orphan_api_v1_orphans_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrphanCreate'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrphanRead'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Create Orphan
      tags:
      - orphans
...
  /api/v1/orphans/{orphan_id}:
    delete:
      description: Soft-delete an orphan record. Restricted to org admins.
      operationId: delete_orphan_api_v1_orphans__orphan_id__delete
      parameters:
      - in: path
        name: orphan_id
        required: true
        schema:
          format: uuid
          title: Orphan Id
          type: string
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Delete Orphan
      tags:
      - orphans
    get:
      operationId: get_orphan_api_v1_orphans__orphan_id__get
      parameters:
      - in: path
        name: orphan_id
        required: true
        schema:
          format: uuid
          title: Orphan Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrphanRead'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Get Orphan
      tags:
      - orphans
    patch:
      operationId: update_orphan_api_v1_orphans__orphan_id__patch
      parameters:
      - in: path
        name: orphan_id
        required: true
        schema:
          format: uuid
          title: Orphan Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrphanUpdate'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrphanRead'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Update Orphan
      tags:
      - orphans
...
  /api/v1/public/orphans:
    get:
      description: 'Anonymous-safe orphan listing. Always filters to case statuses a

        visitor is allowed to sponsor; never exposes archived / deceased /

        pending_review rows.'
      operationId: public_list_orphans_api_v1_public_orphans_get
      parameters:
      - in: query
        name: limit
        required: false
        schema:
          default: 20
          maximum: 50
          minimum: 1
          title: Limit
          type: integer
      - in: query
        name: offset
        required: false
        schema:
          default: 0
          minimum: 0
          title: Offset
          type: integer
      - in: query
        name: country
        required: false
        schema:
          anyOf:
          - maxLength: 2
            minLength: 2
            type: string
          - type: 'null'
          title: Country
      - in: query
        name: gender
        required: false
        schema:
          anyOf:
          - enum:
            - M
            - F
            type: string
          - type: 'null'
          title: Gender
      - in: query
        name: min_age
        required: false
        schema:
          anyOf:
          - maximum: 25
            minimum: 0
            type: integer
          - type: 'null'
          title: Min Age
      - in: query
        name: max_age
        required: false
        schema:
          anyOf:
          - maximum: 25
            minimum: 0
            type: integer
          - type: 'null'
          title: Max Age
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicPage'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Public List Orphans
      tags:
      - public
  /api/v1/public/orphans/{code}:
    get:
      operationId: public_orphan_detail_api_v1_public_orphans__code__get
      parameters:
      - in: path
        name: code
        required: true
        schema:
          title: Code
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicOrphanDetail'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Public Orphan Detail
      tags:
      - public
...